Changing the parameter order in convolution benchmark. Adding inputs as double.

+++ sandbox/apps/python/cnn/conv/init.py
@@ -24,11 +24,11 @@ def init_images(app_data):
     # convert to float image
     IN = np.array(weights)
     IN1 = np.array(input_mat)
-    IN = IN.astype(np.float32).ravel()
-    IN1 = IN1.astype(np.float32).ravel()
+    IN = IN.astype(np.float64).ravel()
+    IN1 = IN1.astype(np.float64).ravel()
 
     # final output image
-    OUT = np.zeros((X, Y, K), np.float32).ravel()
+    OUT = np.zeros((X, Y, K), np.float64).ravel()
 
     img_data = {}
     img_data['IN'] = IN
